different trasport

+++ object-store-core/src/main/java/collaboratory/storage/object/store/core/model/Part.java
@@ -27,7 +27,7 @@ import lombok.NoArgsConstructor;
 final public class Part {
 
   int partNumber;
-  long partSize;
+  int partSize;
   long offset;
   String url;
 }
